builtin.list reduction mathics.eval expansion..
* Move mathics.builtin.list._IterationFunction to mathics.base.IterationFunction
* Move eval-like routines out of mathics.builtin.arithmetic into mathics.eval.numbers
* Move scoping routine out of builtin.base into mathics.eval.scoping
